+++ src/Bridge/Symfony/Blacklister/Not2xxBlacklister.php
@@ -1,0 +1,24 @@
+<?php
+
+/*
+ * This file is part of the SeoOverride project.
+ *
+ * (c) JoliCode <coucou@jolicode.com>
+ *
+ * For the full copyright and license information, please view the LICENSE
+ * file that was distributed with this source code.
+ */
+
+namespace Joli\SeoOverride\Bridge\Symfony\Blacklister;
+
+use Joli\SeoOverride\Bridge\Symfony\Blacklister;
+use Symfony\Component\HttpFoundation\Request;
+use Symfony\Component\HttpFoundation\Response;
+
+class Not2xxBlacklister implements Blacklister
+{
+    public function isBlacklisted(Request $request, Response $response): bool
+    {
+        return $response->getStatusCode() >= 300;
+    }
+}