+++ src/Bridge/Symfony/Blacklister/PathBlacklister.php
@@ -1,0 +1,34 @@
+<?php
+
+/*
+ * This file is part of the SeoOverride project.
+ *
+ * (c) JoliCode <coucou@jolicode.com>
+ *
+ * For the full copyright and license information, please view the LICENSE
+ * file that was distributed with this source code.
+ */
+
+namespace Joli\SeoOverride\Bridge\Symfony\Blacklister;
+
+use Joli\SeoOverride\Bridge\Symfony\Blacklister;
+use Symfony\Component\HttpFoundation\Request;
+use Symfony\Component\HttpFoundation\Response;
+
+class PathBlacklister implements Blacklister
+{
+    /**
+     * @var string
+     */
+    private $pattern;
+
+    public function __construct(string $pattern)
+    {
+        $this->pattern = $pattern;
+    }
+
+    public function isBlacklisted(Request $request, Response $response): bool
+    {
+        return preg_match('#'.$this->pattern.'#i', $request->getPathInfo());
+    }
+}