+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Not yet released
 
 * [BC BREAK] Remove SeoManagerInterface
+* Add blacklist behaviour to avoid useless fetcher runs
 
 ## 0.4.0 (2017-09-12)